% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_orthologs.R
\name{check_arg}
\alias{check_arg}
\title{Check if all elements in the vecter is allowed.}
\usage{
check_arg(vec, allowed)
}
\value{
The vector we are checking.
}
\description{
Check if all elements in the vecter is allowed.
}
\keyword{internal}
