% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_symbols.R
\name{convert_to_entrezID}
\alias{convert_to_entrezID}
\title{Convert symbols to entrezID}
\usage{
convert_to_entrezID(symbols)
}
\arguments{
\item{symbols}{Gene Symbols}
}
\value{
A vector of mapped entrez ID
}
\description{
Convert symbols to entrezID
}
\details{
This function will return a vector of the same length as
the input. If the input doesn't mtach, will return \code{NA} instead
}
\seealso{
\code{\link[AnnotationDbi:AnnotationDb-class]{AnnotationDbi::mapIds()}} and \link[org.Hs.eg.db:org.Hs.egBASE]{org.Hs.eg.db::org.Hs.eg.db}
}
\keyword{internal}
