% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_orthologs.R
\name{find_orthologs}
\alias{find_orthologs}
\title{Find orthologs between species}
\usage{
find_orthologs(genes, from = 9606, to = 7227, filter = "none", version = 8)
}
\arguments{
\item{genes}{A single entrez gene id}

\item{from}{Source species. See @details}

\item{to}{Target species. See @details}

\item{filter}{Filter of the results. See @details.}

\item{version}{What version should be used. Currently v9 is still in beta.}
}
\value{
A tibble with requested information.
}
\description{
Convert entrez gene ID from one species to another using the
\href{https://www.flyrnai.org/diopt}{DIOPT} service.
This function is a wrapper of the DIOPT
\href{https://www.flyrnai.org/tools/diopt/web/api}{API}.
}
\details{
The \code{from} and \code{to} arguments are the NCBI Taxonomy ID.
DIOPT currently supports the following IDs:
\itemize{
\item "all":  All Species (Avaiable for Target only)
\item "4896": Schizosaccharomyces pombe (Fission yeast)
\item "4932": Saccharomyces cerevisiae (Yeast)
\item "6239": Caenorhabditis elegans (Worm)
\item \strong{"7227"}: Drosophila melanogaster (Fly)
\item "7955": Danio rerio (Zebrafish)
\item "8364": Xenopus tropicalis (Western clawed frog)
\item \strong{"9606"}: Homo sapiens (Human)
\item "10090": Mus musculus (Mouse)
\item "3702": Arabidopsis thaliana (Thale cress)
}

If \code{from}/\code{to} is no the same length as the input \code{genes} or 1,
will warning and recycle to match the length of gene.

The \code{filter} argument filter based on score. Options are:
\code{none}/\code{best_match}/\code{exclude_score_less_1}/\code{exclude_score_less_2}.
See \href{https://www.flyrnai.org/tools/diopt/web/api}{API Documentation}
}
