% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_symbols.R
\name{is_entrezID}
\alias{is_entrezID}
\title{Check if an input is enterzID}
\usage{
is_entrezID(input, silent = TRUE)
}
\arguments{
\item{input}{A vector of gene symbol, or name}
}
\value{
Is the input already enterzID
}
\description{
Check if an input is enterzID
}
\keyword{internal}
