% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_orthologs.R
\name{parse_response}
\alias{parse_response}
\title{Function to parse the response.}
\usage{
parse_response(raw_res)
}
\arguments{
\item{raw_res}{The content of a httr response}
}
\description{
Function to parse the response.
}
\keyword{internal}
