% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_orthologs.R
\name{parse_results}
\alias{parse_results}
\title{Helper function to parse the results that has ortholog}
\usage{
parse_results(raw_res)
}
\arguments{
\item{raw_res}{The content of a httr response}
}
\description{
Helper function to parse the results that has ortholog
}
\keyword{internal}
