% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_orthologs.R
\name{parse_search_detail}
\alias{parse_search_detail}
\title{Helper function to parse the query}
\usage{
parse_search_detail(raw_res)
}
\arguments{
\item{raw_res}{The content of a httr response}
}
\description{
Helper function to parse the query
}
\keyword{internal}
