% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_orthologs.R
\name{results_NA_id}
\alias{results_NA_id}
\title{Helper function to parse the results when failed to convert to entrezID}
\usage{
results_NA_id()
}
\description{
Helper function to parse the results when failed to convert to entrezID
}
\keyword{internal}
