% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_orthologs.R
\name{results_no_ortholog}
\alias{results_no_ortholog}
\title{Helper function to parse the results that has no ortholog}
\usage{
results_no_ortholog()
}
\description{
Helper function to parse the results that has no ortholog
}
\keyword{internal}
