% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_orthologs.R
\name{submit_ids}
\alias{submit_ids}
\title{Function to submit a query to DIOPT}
\usage{
submit_ids(id, from = 9606, to = 7227, filter = "none", version = 8)
}
\arguments{
\item{id}{A single entrez gene id}

\item{from}{Source species.}

\item{to}{Target species.}

\item{filter}{Filter of the results, see API \href{https://www.flyrnai.org/tools/diopt/web/api}{API}.}

\item{version}{What version should be used. Currently v9 is still in beta.}
}
\description{
Function to submit a query to DIOPT
}
